I2SPlayback.vhd: FIX - added synchronization to start of audio frame
If for some reason the fifo ever over- or underfills or the input stream
is aborted (incomplete audio frame) then the samples (and L/R) could be
misaligned and produce scrambled sound.

Use the 9th fifo bit to signal the beginning of an audio frame
and let the I2S playback (re-)sync to that when necessary.