Usb2PktProc.vhd: FIX - must not handle other transactions if bufInpVld or bufInpPart
While the buffer is in use (unlikely case) we must delay all other endpoints
until the IN transaction is successfully retried (cannot use the buffer for
other transactions).

Also: added a timeout in case all retries are unsuccessful; we mark the
endpoint as 'dead' if this timeout occurs and STALL until reset but
allow others to use the buffer.