Usb2PktProc.vhd, Usb2PktProcTb.vhd: BUGFIX - EOP handling of full-sized packet in unframed mode
If an IN transaction in unframed mode (bulk/irq) finds "epIb.mstInp.vld = '0'"
and the previous transaction had transferred a full-sized packet then a ZLP
must be sent.

Otherwise the host will keep polling IN because it assumes the transaction
is ongoing. "vld = '0'" corresponds to don='1' in framed mode.

The scenario is equivalent; in framed mode the endpoint must not pause
but has to be able to supply a full transaction which is terminated by 'don'.
If this aligns with a packet boundary then 'don' causes a ZLP.

-> In unframed mode, if 'vld' is withdrawn after a maxPktSize boundary
then a ZLP must be sent.