Usb2FSLSTx.vhd: FIXED 'stp' handling
'stp' must be examined during the cycle following 'nxt' regardless
of bit-stuffing! The old logic used a 'phase' to examine 'stp' but
this might be delayed by bit-stuffing thus missing the actual
'stp' cycle.

Examine 'stp' just after 'nxt' and register the result. Because
a delay may or may not happen we then use 'v.stp' in our logic...