CDCNCM OUT endpoint: changed/FIXED CRC mode
The 'set CRC' bit in the NDP indicates whether the transmitter
must pad the packet if necessary and append a CRC.
The old behaviour stripped a CRC when this bit was set which
seems wrong.

We route this info out to the top so the transmitter/MAC
can be notified.