CDCNCM endpoint: added abort signals
These ports can be used to kill a packet in the INP FIFO
(e.g., when a CRC error is detected) by rewinding the write
pointer.

Likewise, a partial packet in the OUT FIFO can be aborted
e.g., when a collision is detected (by rewinding the read
pointer) and re-played later.