Usb2Pkg.vhd, Usb2MuxEpCtlPkg.vhd, Usb2EpGenericCtlPkg.vhd: new types
Created new subtypes for control request type and recipient. Should
make it easier to identify these bits.

Also: we currently ony use 2 bits of the recipient and ignore the
reserved ones. This change makes it easier to upgrade should it
ever be necessary.

The MuxEpCtlPkg has a new function for initializing config elements
which makes it easier to add new fields and maintain defaults etc...