Reworked/fixed Usb2MuxEpCtl
The mux had only switched the data connection(s) to the
selected agent but failed to communicate that a particular
agent *was* selected (so the agent had to filter requests
consistent with what the mux did).

We now mux the Usb2CtlReqParamType.vld bit so that
when an agent sees this asserted they know that the
mux had actually selected them.