UlpiIO.vhd: assert 'txVld' during the same cycle a valid request is detected
Reduce any delays as much as possible. There is always a window
when 'dir' is not seen by us while it makes it's way through the
input register and OTOH, the phy does not see our request yet as it
makes it through the output registers. Once a collision is detected
we must abort; thus it is desirable to get our request out ASAP.