Usb2DescPkg.vhd (et.al.): removed usage of Usb2AppCfgPkg
Nothing in core/endpoint/Usb2ExampleDev should directly require
Usb2AppCfgPkg. It was only included in Usb2DescPkg so that
the range of Usb2DescIdxType could be restricted. This type
is normally only used in non-synthesized code, so making it
a 'natural' is OK.

Only the Usb2StdCtlEp had synthesized a few signals of
Usb2DescIdxType -- the implementation now uses its own
subtype with a range restricted to DESCRIPTORS_G'range.