Usb2FifoEp.vhd: no need to delay incrementing numFramesInp
The correct numFramesInp is transmitted simultaneously with
the write pointer that accounts for the 'don' flag being written.

So there is no way of the reader detecting 'start of frame'
but getting 'fifoEmpty' because the writePointer has not arrived
on the reader side. The write pointer arrives at the same time
as the number of frames counter.